增加 addJobs updateJobs deleteJobs

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/controller/JobInfoController.java
@@ -55,6 +55,12 @@ public class JobInfoController {
 	@Value("${emaily.access_token.value}")
 	private String emailyAccessTokenValue;
 
+	@Value("${apn.access_token.key}")
+	private String apnAccessTokenKey;
+
+	@Value("${apn.access_token.value}")
+	private String apnAccessTokenValue;
+
 	@Resource
 	private XxlJobInfoDao xxlJobInfoDao;
 
@@ -295,7 +301,12 @@ public class JobInfoController {
 
 	@RequestMapping("/add-jobs")
 	@ResponseBody
-	public ReturnT<Map<String, Integer>> addJobs(@RequestBody List<XxlExtendJobInfo> jobInfoList) {
+	@PermissionLimit(limit = false)
+	public ReturnT<Map<String, Integer>> addJobs(HttpServletRequest request, @RequestBody List<XxlExtendJobInfo> jobInfoList) {
+		if (!request.getHeader(apnAccessTokenKey).equals(apnAccessTokenValue)) {
+			logger.error("[XXL-JOB-ADMIN: addJobs @-1] apn access token is wrong!");
+			return new ReturnT<>(ReturnT.FAIL_CODE, "apn access token is wrong!");
+		}
 		Map<String, Integer> successfulJobIds = new HashMap<>(16);
 		jobInfoList.forEach(jobInfo -> {
 			ReturnT<String> jobReturn = xxlJobService.add(jobInfo);
@@ -308,16 +319,62 @@ public class JobInfoController {
 
 	@RequestMapping("/delete-jobs")
 	@ResponseBody
-	public ReturnT<String> remove(@RequestBody List<Integer> ids) {
+	@PermissionLimit(limit = false)
+	public ReturnT<String> remove(HttpServletRequest request, @RequestBody List<Integer> ids) {
 		logger.info("[XXL-JOB-ADMIN: delete jobs @-1] delete jobs: {}", ids);
+		if (!request.getHeader(apnAccessTokenKey).equals(apnAccessTokenValue)) {
+			logger.error("[XXL-JOB-ADMIN: addJobs @-1] apn access token is wrong!");
+			return new ReturnT<>(ReturnT.FAIL_CODE, "apn access token is wrong!");
+		}
 		return xxlJobService.deleteJobs(ids);
 	}
 
 	@PostMapping("/update-jobs-by-sendTime")
 	@ResponseBody
-	public ReturnT<String> updateJobsBySendTime(@RequestBody List<XxlJobUpdateBySendTime> xxlJobUpdateBySendTime) {
+	@PermissionLimit(limit = false)
+	public ReturnT<String> updateJobsBySendTime(HttpServletRequest request, @RequestBody List<XxlJobUpdateBySendTime> xxlJobUpdateBySendTime) {
 		logger.info("[XXL-JOB-ADMIN: update jobs by sendTime @-1] param : {}", xxlJobUpdateBySendTime);
+		if (!request.getHeader(apnAccessTokenKey).equals(apnAccessTokenValue)) {
+			logger.error("[XXL-JOB-ADMIN: addJobs @-1] apn access token is wrong!");
+			return new ReturnT<>(ReturnT.FAIL_CODE, "apn access token is wrong!");
+		}
 		return xxlJobService.updateJobsBySendTime(xxlJobUpdateBySendTime);
+	}
+
+	@RequestMapping("/my-add")
+	@ResponseBody
+	@PermissionLimit(limit = false)
+	public ReturnT<String> myAdd(HttpServletRequest request, XxlJobInfo jobInfo) {
+		logger.info("[XXL-JOB-ADMIN: add job @-1] param : {}", jobInfo);
+		if (!request.getHeader(apnAccessTokenKey).equals(apnAccessTokenValue)) {
+			logger.error("[XXL-JOB-ADMIN: myAdd @-1] apn access token is wrong!");
+			return new ReturnT<>(ReturnT.FAIL_CODE, "apn access token is wrong!");
+		}
+		return xxlJobService.add(jobInfo);
+	}
+
+	@RequestMapping("/my-update")
+	@ResponseBody
+	@PermissionLimit(limit = false)
+	public ReturnT<String> myUpdate(HttpServletRequest request, XxlJobInfo jobInfo) {
+		logger.info("[XXL-JOB-ADMIN: update job @-1] param : {}", jobInfo);
+		if (!request.getHeader(apnAccessTokenKey).equals(apnAccessTokenValue)) {
+			logger.error("[XXL-JOB-ADMIN: myUpdate @-1] apn access token is wrong!");
+			return new ReturnT<>(ReturnT.FAIL_CODE, "apn access token is wrong!");
+		}
+		return xxlJobService.update(jobInfo);
+	}
+
+	@RequestMapping("/my-remove")
+	@ResponseBody
+	@PermissionLimit(limit = false)
+	public ReturnT<String> myRemove(HttpServletRequest request, int id) {
+		logger.info("[XXL-JOB-ADMIN: remove job @-1] id : {}", id);
+		if (!request.getHeader(apnAccessTokenKey).equals(apnAccessTokenValue)) {
+			logger.error("[XXL-JOB-ADMIN: myRemove @-1] apn access token is wrong!");
+			return new ReturnT<>(ReturnT.FAIL_CODE, "apn access token is wrong!");
+		}
+		return xxlJobService.remove(id);
 	}
 
 }
